Refactor: 카메라 컴포넌트 폴더구조 변경 및 에러, 로딩 페이지 분리 외 수정
