+++ src/app/dev/camera/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { MirrorView, useCamera, VideoView } from '@/components/camera';
-import { SpinnerIcon } from '@/components/icon';
+import CameraErrorView from '@/components/camera/components/CameraErrorView';
 
 const CamerExample = () => {
   const {
@@ -14,23 +14,8 @@ const CamerExample = () => {
     error,
   } = useCamera();
 
-  if (isLoading) {
-    return (
-      <div className="flex h-screen w-screen items-center justify-center">
-        <SpinnerIcon />
-      </div>
-    );
-  }
-
   if (error) {
-    return (
-      <div className="flex-col items-center justify-center">
-        <div className="flex items-center justify-center">
-          에러가 발생했습니다. 다시 시도해주세요.
-        </div>
-        <div className="flex items-center justify-center">ERROR: {error}</div>
-      </div>
-    );
+    return <CameraErrorView error={error} />;
   }
 
   return (
@@ -57,9 +42,18 @@ const CamerExample = () => {
 
       <div>
         {isRecording ? (
-          <MirrorView stream={previewStream} />
+          <MirrorView
+            stream={previewStream}
+            isLoading={isLoading}
+            error={error}
+          />
         ) : (
-          mediaBlobUrl && <VideoView videoUrl={mediaBlobUrl} />
+          mediaBlobUrl && (
+            <VideoView
+              videoUrl={mediaBlobUrl}
+              error={error}
+            />
+          )
         )}
       </div>
     </div>
