+++ src/components/camera/index.ts
@@ -1,3 +1,3 @@
-export { default as MirrorView } from './MirrorView';
-export { default as useCamera } from './useCamera';
-export { default as VideoView } from './VideoView';
+export { default as MirrorView } from './components/MirrorView';
+export { default as VideoView } from './components/VideoView';
+export { default as useCamera } from './hook/useCamera';
