+++ src/components/camera/style/camera.css
